UPDATED: - 인터페이스 => types 디렉터리로 이동
