+++ src/helpers/ChangePasswordValidator.ts
@@ -1,6 +1,7 @@
 import { FormValidationResult } from './FormValidationResult';
 import { FormValidator } from './FormValidator';
-import { IChangePasswordFormValues } from './IChangePasswordFormValues';
+import { IChangePasswordFormValues } from '../typings/IValidatorInputValues';
+
 export class ChangePasswordValidator extends FormValidator {
     public validate(formData: IChangePasswordFormValues) {
         const result: FormValidationResult = new FormValidationResult([
