--- src/helpers/ICategoriesFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface ICategoriesFormValue extends IDictionary<any> {
-    categories: any[];
-}
