--- src/helpers/ICategoryNameFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface ICategoryNameFormValue extends IDictionary<any> {
-    name: string;
-}
