--- src/helpers/ICategoryOrdinalFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface ICategoryOrdinalFormValue extends IDictionary<any> {
-    ordinal?: number;
-}
