--- src/helpers/ICategorySlugFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface ICategorySlugFormValue extends IDictionary<any> {
-    slug?: string;
-}
