--- src/helpers/IChangePasswordFormValues.ts
@@ -1,9 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-import { IPasswordFormValue } from './IPasswordFormValue';
-import { IPasswordConfirmFormValue } from './IPasswordConfirmFormValue';
-import { ICurrentPasswordFormValue } from './ICurrentPasswordFormValue';
-export interface IChangePasswordFormValues
-    extends IPasswordFormValue,
-        IPasswordConfirmFormValue,
-        ICurrentPasswordFormValue,
-        IDictionary<any> {}
