--- src/helpers/ICurrentPasswordFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface ICurrentPasswordFormValue extends IDictionary<any> {
-    currentPassword: string;
-}
