--- src/helpers/IDisplayNameFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface IDisplayNameFormValue extends IDictionary<any> {
-    displayName: string;
-}
