--- src/helpers/IEmailFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface IEmailFormValue extends IDictionary<any> {
-    email: string;
-}
