--- src/helpers/IFormValidationResult.ts
@@ -1,4 +1,0 @@
-export interface IFormValidationResult {
-    valid: boolean;
-    messages: string[];
-}
