--- src/helpers/IMarkdownFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface IMarkdownFormValue extends IDictionary<any> {
-    markdown: string;
-}
