--- src/helpers/IPasswordConfirmFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface IPasswordConfirmFormValue extends IDictionary<any> {
-    passwordConfirm: string;
-}
