--- src/helpers/IPasswordFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface IPasswordFormValue extends IDictionary<any> {
-    password: string;
-}
