--- src/helpers/ITagsFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface ITagsFormValue extends IDictionary<any> {
-    tags?: any[];
-}
