--- src/helpers/ITitleFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface ITitleFormValue extends IDictionary<any> {
-    title: string;
-}
