--- src/helpers/IUsernameFormValue.ts
@@ -1,4 +1,0 @@
-import { IDictionary } from '../typings/IDictionary';
-export interface IUsernameFormValue extends IDictionary<any> {
-    username: string;
-}
