+++ src/helpers/SignUpFormValidator.ts
@@ -1,16 +1,21 @@
-import { IFormValidationResult } from './IFormValidationResult';
 import { FormValidationResult } from './FormValidationResult';
-import { ISignupFormValues } from './ISignupFormValues';
 import { FormValidator } from './FormValidator';
+import {
+    ISignupFormValues,
+    IFormValidationResult,
+} from '../typings/IValidatorInputValues';
+
 export class SignUpFormValidator extends FormValidator {
     public validate(formValues: ISignupFormValues): IFormValidationResult {
-        const validationResult: IFormValidationResult = new FormValidationResult([
-            this.checkEmail(formValues),
-            this.checkPassword(formValues),
-            this.checkPasswordConfirm(formValues),
-            this.checkUsername(formValues),
-            this.checkDisplayName(formValues),
-        ]);
+        const validationResult: IFormValidationResult = new FormValidationResult(
+            [
+                this.checkEmail(formValues),
+                this.checkPassword(formValues),
+                this.checkPasswordConfirm(formValues),
+                this.checkUsername(formValues),
+                this.checkDisplayName(formValues),
+            ],
+        );
         return validationResult;
     }
 }
