+++ src/typings/IValidationResult.ts
